Adds alignBlocksViaPortHandle_gcb.m to the repository
Adds the function "alignBlocksViaPortHandle_gcb.m", the first working example of the repository.

The for a given selected block inside a model, the function will align this block with its connected reference block by using the Port Handle Position property of the reference block, so that their connection line becomes a "straight line".